Update and rename lyrics_view_window.py to get_lyrics.py
